Feat: specify custom temperature of the system
The default temperature of the system used to be 2.7 K, a low estimate
for the temperature of cosmic background radiation. This commit allows
specifying a different temperature, keeping 2.7 K as a default in case
none is provided. MOAC uses temperatures to calculate the Landauer
limit.

The CLI and docs were updated to include this feature.

Updated the README example to utilize this feature: include a more
realistic temperature estimate of the Earth's internal temperature.