Test: custom charsets shouldn't bloat pw length
pwgen_test now ensures that the number of generated passwords with
excessive lengths isn't too high more than a negligable amount.

Only one set of paramsmakes this possible, so this is only allowed for
that situation and only a negligible number of times.

Add a new set of params for test cases that used to cause this error.
The bug was fixed in ee6732768b1ddfce03b45ee24ddd998b896c7a48.

Resolves https://todo.sr.ht/~seirdy/MOAC/13