Fix: correctly subset charsets, with tests
When charsets are subsets of each other, they should be shrunk to
minimize the overlap while keeping all the unique characters.

New behavior:

Say charsets A and B are completely distinct, with no overlap. Say
charset C is a combination of A and B.

If GenPW is told to generate a password using charsets A and C, it
should replace C with B.

Say A is actually a combination of two charsets, a1 and a2. Charset D is
a combination of a1 and B.

If GenPW is told to generate a password using charsets A and D, then it
should generate a password containing charsets a2 and B instead.

The above rules guarantee that each charset's unique entries gets
represented in the final passwords.

This should also reduce the likelihood of an over-inflated password
entropy that's caused by overlapping charsets.

Resolves https://todo.sr.ht/~seirdy/MOAC/14