Feat!: validate: avoid impossible temperatures
BREAKING CHANGE: moac.MinEntropy returns (float64, error).

Ensure that given temperatures are physically possible, between 0 and
the Planck Temperature (non-inclusive).

This introduces a new internal package, internal/bounds, to hold any
limits on given values. At ths moment it only holds values for
temperature, but it will grow in future commits.

This changes the function signature for MinEntropy to support error
returning, since it can now return an error when given an impossible
temperature. That required updating many other files to reflect this
change and handle the new error retval.

Updating the CLI to reflect that change resulted in duplicate code
for exiting on an error crossing a threshold. Solve this by creating
"internal/cli" for shared functions between cmd/{moac,moac-pwgen}, and
put "ExitOnErr" in it.

Acceptable limits for temperature are now documented in the manpages.