Refactor! adopt Charsets interface/subpkg (1/3)
BREAKING CHANGE: move the exported charsets out of the entropy package
and into their own sub-package.

Create a new sub-package "charsets" that holds the default charsets, now
as exported consts. The charsets are collected into a global
DefaultCharsets variable; unlike the previous Charsets variable, this is
not a map[string][]rune, but a new CharsetCollection interface
implementation.

This is the first half of a multi-commit change that will streamline the
process of dealing with predefined and custom charsets.

Note: This stage of the refactor demanded some updates to the linter
config, obsoleting some //nolint comments.