Fix: improve minimization of charset redundancy
Tweak charsets.CustomCharset.Add's approach to removing redundant
entries in charsets to improve performance and fix a bug that allowed
unsorted charsets.

Before: redundancy elimination of an old and new charset's redundancies
would involve separating the overlap from old/new charsets, adding it to
the shorter charset, and sorting the result. This was expensive and
resulted in unsorted outputs.

After: redundancy elimination directly alters the longer of the two
charsets, ensuring charsets don't need expensive re-sorting. Adding a
charset also exits early if the new charset is reduced to an empty set.