Build: fix Makefile portability/conventions
Backticks only work in bmake, but fail in GNU Make. This commit replaces
them with "!=" assignments which are portable across both, as of GNU
Make v4.0.

This also replaces unnecessary ?= with =.

Finally, this removes the unnecessary and problematic .POSIX declaration
at the top; that sets a bunch of unnecessary vars.