Refactor: consistency across cmd/* parseOpts()
Reflect recent changes to cmd/moac/main.go's parseOpts() in moac-pwgen;
this will help with the upcoming CLI testing.

These changes include returning an error in ParseOpts instead of exiting
early. Since that error is the same for both moac and moac-pwgen, move
its declaration into internal/cli/cli.go so it can be shared between the
two.

Trivial fix:
In the process, I noticed that one of the parameters in genpw_test was
set to allow some improbable false-negatives slip through; fix that.