Doc: mention transition from PGP to Signify
Signify is a much simpler and more secure alternative to PGP for
cryptographic signatures; document a transition from PGP sigs to
Minisign-generated Signify sigs for releases, and that commits will be
signed with both (for now...).

See https://codeberg.org/SkyFox42/git-minisign for more info on signing
Git commits with minisign.