Fix: make GenPW validate password requirements
GenPW validates PwRequirements in a way similar to moac.Givens. It
ensures that lengths and entropies are above zero, and that GenPW has at
least two unique characters to pick from.

According to MOAC's entropy algorithm, passwords with a charset size of
one have an entropy of zero; this messes up password generation since
password entropy is constant, regardless of length. The entropy
measurement is correct, but we can't generate passwords under these
conditions: treating the situation as invalid seems reasonable.

Added regression tests for all parts of the new validation, and document
requirements for password constraints in the moac-pwgen manpage and
the PwRequirements godoc.

Resolves https://todo.sr.ht/~seirdy/MOAC/28
Resolves https://todo.sr.ht/~seirdy/MOAC/29