Tests: more testcases + remove ineffective lines
I had some fun with gobco and go-mutesting for branch coverage and
mutation testing, respectively. They revealed that some branches weren't
covered despite good statement coverage. Add some test cases for those
situations, some of which are borrowed from testscripts.

This also revealed that the IsDefault() checks in pwgen were completely
unnecessary; remove them and deprecate IsDefault().

I noticed that mem usage on tests crossed a little threshold with the
addition of additional testcases. Fix that by replacing some hugeparams
and range-value-copying with pointers and indexing.

The use of gobco and go-mutesting demonstrated that statement coverage
isn't enough, and that tests need to do more than execute a line to show
that it's necessary + correct. Document this in the CONTRIBUTING.md's
section on testing standards.