Test: add test case for custom guesses-per-second
Since I'm about to add functionality to the CLI to specify a custom
number of guesses per second, it seemed appropriate to test this
functionality in the library first. Surprisingly, a custom
guesses-per-second was never accounted for until now.

When I add the ability to specify custom guesses-per-second on the CLI,
the corresponding scenario test can re-use these values.