Refactor: remove impossible/unnecessary panics
Remove impossible and unnecessary panics in givens.go; these checked for
impossible situations that are now accounted for and tested. For
instance, one check ensured that energy was never zero, but since then
givens is guaranteed to set energy to a nonzero value or error
appropriately.

Re-ordered the functions to follow a more logical flow.